MCP_application/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-identity  
azure-search-documents==11.6.0b6
extract-msg  
openai  
pydantic  
python-dotenv  
requests  
azure.cosmos
streamlit
fastmcp
semantic-kernel
mcp